Refactor chata_stream
