+++ backend/app/api/v1/schemas/__init__.py
@@ -1,0 +1,5 @@
+"""API schemas for chat endpoints."""
+
+from app.api.v1.schemas.chat_schemas import ChatMessage, MessagePart, StreamRequest
+
+__all__ = ["MessagePart", "ChatMessage", "StreamRequest"]
