+++ backend/app/api/v1/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities for API v1 endpoints."""
